Add application using tess-two
This application is targeting receipt image reading. It can read text
and arrange by words. The goal is to itemize a receipt using the box
locations. Currently this works on clear white background receipts. For
darker receipts there is code to read the receipt, but will need to be
uncomments (resulting in harder time reading clear receipts).